+++ src/Form/index.test.js
@@ -13,25 +13,6 @@ class DummyInput extends React.Component {
   }
 }
 
-// since simpleSchema comes from Meteor, we need to make a stub.
-class SchemaStub {
-  constructor (schema) {
-    this._schema = schema
-  }
-
-  schema (key) {
-    return this._schema[key] || {}
-  }
-
-  label (key) {
-    return `label-for-${key}`
-  }
-
-  newContext () {
-    return this
-  }
-}
-
 test('Should render by default a <form>', () => {
   const component = shallow(
     <Form>
@@ -94,45 +75,4 @@ test('should render the form correctly', () => {
   it('should render the <DummyInput>', () => {
     expect(component.find('DummyInput').length).toBe(1)
   })
-})
-
-test('should generate input keys correctly based on the schema', () => {
-  const component = mount(
-    <Form
-      type='insert'
-      schema={new SchemaStub({
-        name: {
-          type: String,
-          srf: {
-            type: DummyInput
-          },
-          label: 'foobarONE'
-        },
-        ignoreMeSenpai: {
-          type: Number,
-          srf: {
-            type: DummyInput,
-            omit: true
-          },
-          label: 'barfooTWO'
-        }
-      })}
-    >
-      <Field fieldName='name' type={DummyInput} />
-      <Field fieldName='deviceId' type={DummyInput} />
-    </Form>
-  )
-
-  const generateInputsForKeys = component.instance().generateInputsForKeys.bind(component.instance())
-
-  let nameComponent = generateInputsForKeys(['name'])[0]
-  expect(nameComponent.key).toEqual('name')
-  expect(nameComponent.props.fieldName).toEqual('name')
-
-  let nameComponentWithParent = generateInputsForKeys(['name'], 'i.am.a.parent')[0]
-  expect(nameComponentWithParent.key).toEqual('i.am.a.parent.name')
-  expect(nameComponent.props.fieldName).toEqual('name')
-
-  let omitComponent = generateInputsForKeys(['ignoreMeSenpai'])
-  expect(omitComponent.length).toEqual(0)
 })